Implement FCM for receving messages on bare workflow
Sending notifications from expo api works well. The first try after the app got build it didn't work, but the second time it did. We shall look in the reason behind that by logging and catching errors.

Changing the default notification icon wasn't successful. It may take a while to find the cause.